perf(e2e-storm): migrate harness off dead .mcp.json discovery to daemon /rpc
The MCP→CLI cutover (2651ade78, fab76e7d4, 15595a854) removed the in-process
MCP server and its `.mcp.json` handshake; the unified daemon now publishes
`.voicetree/{rpc.port,auth-token}` and serves tools over JSON-RPC at `/rpc`.
The e2e-storm-mvp harness still polled `.mcp.json` + POSTed MCP
initialize/tools-call to `/mcp`, so it timed out against any current build.

- launchElectron: discover via @vt/vt-rpc createRpcClientForProject (rpc.port +
  auth-token), gated on a list_agents readiness probe; return a DaemonRpcClient
  instead of an mcpPort. Strips an ambient $VOICETREE_DAEMON_URL so discovery
  binds strictly to THIS electron's project daemon.
- runFakeAgent: drive spawn_agent/list_agents through DaemonRpcClient.call
  (same transport the vt CLI uses). The daemon unwraps tool responses, so
  success returns the payload directly and isError surfaces as a JSON-RPC
  error whose data carries the failure sentence (preserved in spawnError).
- index/report: rename mcpPort→daemonClient/daemonUrl vocabulary, drop the MCP
  initialize handshake.

Also fixes botched-rename casualties from the path-vocabulary refactors
(5f0b28fb8, 3965665fc): duplicate `projectDir` identifiers that left the file
non-compiling. Decomposes the 919-line index.ts (over the repo's 500-line
limit) into functional sibling modules: stormStats (pure timing/output stats),
runConfig (args/run-context/settings), captures (CDP metrics/profile/screenshot
samplers). Baseline load default is now 50 agents × 8 nodes (400 nodes, the
>300-node regime where perf degrades).
